+++ crates/maudio/src/device.rs
@@ -21,6 +21,7 @@ use crate::{
         device_state::DeviceState,
         device_type::DeviceType,
     },
+    logging::{LogRef, StoredLogs},
     pcm_frames::PcmFormat,
     util::{device_notif::DeviceStateNotifier, proc_notif::ProcFramesNotif},
     AllocationCallbacks, Binding, MaResult,
@@ -38,12 +39,14 @@ pub mod device_type;
 /// Manages the lifetime of a `ma_device` and provides control over
 /// playback, capture, and device state.
 pub struct Device<F: PcmFormat> {
-    pub(crate) inner: Arc<DeviceInner<F>>,
+    pub(crate) inner: Arc<DeviceInner>,
+    _format: PhantomData<F>,
     // Device cannot be sync.
     _not_sync: PhantomData<Cell<()>>,
 }
 
-pub(crate) struct DeviceInner<F: PcmFormat> {
+#[doc(hidden)]
+pub struct DeviceInner {
     inner: *mut sys::ma_device,
     _playback_device_id: Option<DeviceId>, // Ref count. Needs to be kept alive.
     _capture_device_id: Option<DeviceId>,  // Ref count. Needs to be kept alive.
@@ -52,10 +55,10 @@ pub(crate) struct DeviceInner<F: PcmFormat> {
     callback_panic: Arc<AtomicBool>,       // true = callback panicked and is now poisoned
     callback_process_notifier: ProcFramesNotif,
     state_notifier: Option<DeviceStateNotifier>, // used by ma_device_notification
-    _format: PhantomData<F>,
+    pub(crate) logs: StoredLogs,
 }
 
-impl<F: PcmFormat> Binding for DeviceInner<F> {
+impl Binding for DeviceInner {
     type Raw = *mut sys::ma_device;
 
     fn to_raw(&self) -> Self::Raw {
@@ -64,8 +67,8 @@ impl<F: PcmFormat> Binding for DeviceInner<F> {
 }
 
 // Required for Arc<DeviceInner> to implement Send
-unsafe impl<F: PcmFormat> Send for DeviceInner<F> {}
-unsafe impl<F: PcmFormat> Sync for DeviceInner<F> {}
+unsafe impl Send for DeviceInner {}
+unsafe impl Sync for DeviceInner {}
 
 impl<F: PcmFormat> Binding for Device<F> {
     type Raw = *mut sys::ma_device;
@@ -293,6 +296,10 @@ impl<F: PcmFormat> Device<F> {
         device_ffi::ma_device_stop(self)
     }
 
+    pub fn log(&self) -> LogRef {
+        device_ffi::ma_device_get_log(self)
+    }
+
     /// Returns `true` if the data callback previously panicked.
     ///
     /// When this happens, the callback is considered poisoned and will no longer run.
@@ -356,8 +363,9 @@ impl<F: PcmFormat> Device<F> {
                 callback_panic: cb_info.data_callback_panic,
                 callback_process_notifier: data_notif,
                 state_notifier: Some(cb_info.state_notif.clone()),
-                _format: PhantomData,
+                logs: StoredLogs::default(),
             }),
+            _format: PhantomData,
             _not_sync: PhantomData,
         })
     }
@@ -379,6 +387,7 @@ pub(crate) mod device_ffi {
             device_type::DeviceType,
             private_device, AsDevicePtr, Device, DeviceInner,
         },
+        logging::{LogOwner, LogRef},
         pcm_frames::PcmFormat,
         AsRawRef, Binding, MaResult, MaudioError,
     };
@@ -425,7 +434,7 @@ pub(crate) mod device_ffi {
         MaudioError::check(res)
     }
 
-    pub fn ma_device_uninit<F: PcmFormat>(device: &mut DeviceInner<F>) {
+    pub fn ma_device_uninit(device: &mut DeviceInner) {
         unsafe { sys::ma_device_uninit(device.to_raw()) };
     }
 
@@ -442,15 +451,15 @@ pub(crate) mod device_ffi {
         }
     }
 
-    // TODO: Implement log
+    // Callback: not safe
+    // Theadsafe: not safe
     #[inline]
-    #[allow(dead_code)]
-    pub fn ma_device_get_log<D: AsDevicePtr + ?Sized>(context: &D) -> Option<*mut sys::ma_log> {
-        let ptr = unsafe { sys::ma_device_get_log(private_device::device_ptr(context)) };
-        if ptr.is_null() {
-            None
-        } else {
-            Some(ptr)
+    pub fn ma_device_get_log<F: PcmFormat>(device: &Device<F>) -> LogRef {
+        let ptr = unsafe { sys::ma_device_get_log(device.to_raw()) };
+
+        LogRef {
+            inner: ptr,
+            _owner: LogOwner::Device(device.inner.clone()),
         }
     }
 
@@ -632,7 +641,7 @@ pub(crate) mod device_ffi {
     }
 }
 
-impl<F: PcmFormat> Drop for DeviceInner<F> {
+impl Drop for DeviceInner {
     fn drop(&mut self) {
         device_ffi::ma_device_uninit(self);
         (self.callback_user_data_drop)(self.callback_user_data);