+++ crates/maudio/src/engine.rs
@@ -97,6 +97,7 @@ use crate::{
         process_cb::ProcessState,
         resource::{ResourceManager, ResourceManagerRef},
     },
+    logging::{LogInner, LogRef, StoredLogs},
     sound::{
         sound_builder::SoundBuilder,
         sound_ffi,
@@ -133,14 +134,16 @@ pub struct Engine(pub(crate) Arc<EngineInner>);
 #[doc(hidden)]
 pub struct EngineInner {
     inner: *mut sys::ma_engine,
-    _playback_device_id: Option<DeviceId>,  // keep alive
-    _device: Option<Arc<DeviceInner<f32>>>, // keep alive
+    _playback_device_id: Option<DeviceId>, // keep alive
+    _device: Option<Arc<DeviceInner>>,     // keep alive
     _resource_manager: Option<ResourceManager<f32>>, // keep alive
+    _logger: Option<Arc<LogInner>>,        // keep alive
     process_data_ptr: Option<*mut ProcessState>, // userdata (self.inner.pProcessUserData)
     process_data_panic: Option<Arc<AtomicBool>>, // true = callback panicked and is now poisoned
     process_data_notif: Option<ProcFramesNotif>,
     state_notifier: Option<DeviceStateNotifier>,
     reader_exists: Arc<AtomicBool>,
+    pub(crate) logs: StoredLogs,
 }
 
 unsafe impl Send for EngineInner {}
@@ -275,10 +278,11 @@ impl Engine {
     }
 
     fn new_with_config(config: &EngineBuilder) -> MaResult<Self> {
-        let (device, rm, dev_id) = (
+        let (device, rm, dev_id, log) = (
             config.device.clone(),
             config.resource_manager.clone(),
             config.playback_device_id.clone(),
+            config.log.clone(),
         );
         let mut mem: Box<MaybeUninit<sys::ma_engine>> = Box::new(MaybeUninit::uninit());
         engine_ffi::engine_init(config, mem.as_mut_ptr())?;
@@ -289,11 +293,13 @@ impl Engine {
             _playback_device_id: dev_id,
             _device: device,
             _resource_manager: rm,
+            _logger: log,
             process_data_ptr: None,
             process_data_panic: None,
             process_data_notif: None,
             state_notifier: None,
             reader_exists: Arc::new(AtomicBool::new(false)),
+            logs: StoredLogs::default(),
         })))
     }
 
@@ -317,17 +323,28 @@ impl Engine {
             _playback_device_id: config.playback_device_id.take(),
             _device: config.device.take(),
             _resource_manager: config.resource_manager.take(),
+            _logger: config.log.clone(),
             process_data_ptr: config.process_data.process_data_ptr,
             process_data_panic: config.process_data.process_data_panic.take(),
             process_data_notif: data_notif,
             state_notifier: state_notif,
             reader_exists: Arc::new(AtomicBool::new(false)),
+            logs: StoredLogs::default(),
         })))
     }
 
     /// Equivalent to calling [`SoundBuilder::new()`]
     pub fn sound_config<'a, 'b>(&'a self) -> SoundBuilder<'a, 'b> {
         SoundBuilder::init(self)
+    }
+
+    /// Plays a fire and forget sound once without returning a handle.
+    ///
+    /// The engine manages the sound internally for the duration of playback.
+    ///
+    /// Will return an [`ErrorKinds::InvalidCString`](crate::ErrorKinds) error if path is not valid UTF8
+    pub fn play_one_shot(&self, path: &Path) -> MaResult<()> {
+        engine_ffi::ma_engine_play_sound(self, path)
     }
 
     /// Creates an empty sound node with no audio source.
@@ -519,6 +536,10 @@ impl Engine {
         engine_ffi::ma_engine_get_device(self)
     }
 
+    pub fn log(&self) -> LogRef {
+        engine_ffi::ma_engine_get_log(self)
+    }
+
     /// Returns the engine’s **endpoint node**.
     ///
     /// The endpoint node is the final node in the engine’s internal node graph.
@@ -692,6 +713,7 @@ pub(crate) mod engine_ffi {
             resource::{ResourceManagerRef, RmOwner},
             AsEnginePtr, Binding, Engine, EngineInner, EngineReader,
         },
+        logging::{LogOwner, LogRef},
         AsRawRef, MaResult, MaudioError,
     };
 
@@ -706,6 +728,26 @@ pub(crate) mod engine_ffi {
         unsafe {
             sys::ma_engine_uninit(engine.inner);
         }
+    }
+
+    #[inline]
+    pub fn ma_engine_play_sound(engine: &Engine, path: &std::path::Path) -> MaResult<()> {
+        #[cfg(not(windows))]
+        let path = crate::cstring_from_path(path)?;
+
+        #[cfg(windows)]
+        let path = {
+            let path = path
+                .to_str()
+                .ok_or(MaudioError::new_ma_error(crate::ErrorKinds::InvalidCString))?;
+            std::ffi::CString::new(path.as_bytes())
+                .map_err(|_| MaudioError::new_ma_error(crate::ErrorKinds::InvalidCString))?
+        };
+
+        let res = unsafe {
+            sys::ma_engine_play_sound(engine.to_raw(), path.as_ptr(), std::ptr::null_mut())
+        };
+        MaudioError::check(res)
     }
 
     #[inline]
@@ -789,11 +831,14 @@ pub(crate) mod engine_ffi {
         }
     }
 
-    // TODO: Implement Log(Ref?)
     #[inline]
-    #[allow(dead_code)]
-    pub fn ma_engine_get_log(engine: &Engine) -> *mut sys::ma_log {
-        unsafe { sys::ma_engine_get_log(engine.to_raw()) }
+    pub fn ma_engine_get_log(engine: &Engine) -> LogRef {
+        let ptr = unsafe { sys::ma_engine_get_log(engine.to_raw()) };
+
+        LogRef {
+            inner: ptr,
+            _owner: LogOwner::Engine(engine.0.clone()),
+        }
     }
 
     #[inline]