+++ crates/maudio/src/engine/node_graph/nodes.rs
@@ -111,7 +111,6 @@ impl TryFrom<sys::ma_node_state> for NodeState {
 /// See [`NodeBuilder`](crate::engine::node_graph::node_builder)
 pub struct Node<C> {
     pub(crate) inner: *mut NodeInner<C>,
-    _not_sync: PhantomData<Cell<()>>,
 }
 
 #[repr(C)]
@@ -122,11 +121,10 @@ pub(crate) struct NodeInner<C> {
     pub(crate) custom: C,
     pub(crate) op: NodeFunction,
     pub(crate) owner: GraphOwner,
-    pub(crate) _not_sync: PhantomData<Cell<()>>,
 }
 
-unsafe impl<C> Send for NodeInner<C> {}
-unsafe impl<C> Sync for NodeInner<C> {}
+unsafe impl<C> Send for NodeInner<C> where C: Send {}
+unsafe impl<C> Sync for NodeInner<C> where C: Sync {}
 
 impl<C> AsRawRef for Node<C> {
     type Raw = sys::ma_node_base;
@@ -218,7 +216,6 @@ impl<C> Node<C> {
             custom,
             op,
             owner: private_node_graph::clone_owner(node_graph),
-            _not_sync: PhantomData,
         });
 
         let base_ptr = core::ptr::addr_of_mut!(inner.base);
@@ -232,10 +229,7 @@ impl<C> Node<C> {
             inner_ptr.cast::<u8>(),
         );
 
-        Ok(Node {
-            inner: inner_ptr,
-            _not_sync: PhantomData,
-        })
+        Ok(Node { inner: inner_ptr })
     }
 
     pub(crate) fn build_required_frames<N>(
@@ -272,7 +266,6 @@ impl<C> Node<C> {
             custom,
             op,
             owner: private_node_graph::clone_owner(node_graph),
-            _not_sync: PhantomData,
         });
 
         let base_ptr = core::ptr::addr_of_mut!(inner.base);
@@ -286,10 +279,7 @@ impl<C> Node<C> {
             inner_ptr.cast::<u8>(),
         );
 
-        Ok(Node {
-            inner: inner_ptr,
-            _not_sync: PhantomData,
-        })
+        Ok(Node { inner: inner_ptr })
     }
 }
 
@@ -566,21 +556,21 @@ impl<T: AsNodePtr + ?Sized> NodeOps for T {}
 pub trait NodeOps: AsNodePtr {
     /// Attaches `output_bus` of this node to `other_node_input_bus` of `other_node`.
     fn attach_output_bus<P: AsNodePtr + ?Sized>(
-        &mut self,
+        &self,
         output_bus: u32,
-        other_node: &mut P,
+        other_node: &P,
         other_node_input_bus: u32,
     ) -> MaResult<()> {
         node_ffi::ma_node_attach_output_bus(self, output_bus, other_node, other_node_input_bus)
     }
 
     /// Detaches the specified output bus from its connected input bus.
-    fn detach_output_bus(&mut self, output_bus: u32) -> MaResult<()> {
+    fn detach_output_bus(&self, output_bus: u32) -> MaResult<()> {
         node_ffi::ma_node_detach_output_bus(self, output_bus)
     }
 
     /// Detaches all output buses from their connected input buses.
-    fn detach_all_outputs(&mut self) -> MaResult<()> {
+    fn detach_all_outputs(&self) -> MaResult<()> {
         node_ffi::ma_node_detach_all_output_buses(self)
     }
 
@@ -605,12 +595,12 @@ pub trait NodeOps: AsNodePtr {
     }
 
     /// Returns the volume for the given output bus.
-    fn output_bus_volume(&mut self, out_bus_index: u32) -> f32 {
+    fn output_bus_volume(&self, out_bus_index: u32) -> f32 {
         node_ffi::ma_node_get_output_bus_volume(self, out_bus_index)
     }
 
     /// Sets the volume for the given output bus.
-    fn set_output_bus_volume(&mut self, out_bus_index: u32, volume: f32) -> MaResult<()> {
+    fn set_output_bus_volume(&self, out_bus_index: u32, volume: f32) -> MaResult<()> {
         node_ffi::ma_node_set_output_bus_volume(self, out_bus_index, volume)
     }
 
@@ -623,7 +613,7 @@ pub trait NodeOps: AsNodePtr {
         node_ffi::ma_node_get_state(self)
     }
 
-    fn set_state(&mut self, state: NodeState) -> MaResult<()> {
+    fn set_state(&self, state: NodeState) -> MaResult<()> {
         node_ffi::ma_node_set_state(self, state)
     }
 
@@ -633,7 +623,7 @@ pub trait NodeOps: AsNodePtr {
     }
 
     /// Returns the global time (in PCM frames) at which `state` becomes active.
-    fn set_state_time(&mut self, state: NodeState, global_time: u64) -> MaResult<()> {
+    fn set_state_time(&self, state: NodeState, global_time: u64) -> MaResult<()> {
         node_ffi::ma_node_set_state_time(self, state, global_time)
     }
 
@@ -657,7 +647,7 @@ pub trait NodeOps: AsNodePtr {
     }
 
     /// Sets the current local time (in PCM frames) of the node.
-    fn set_time(&mut self, local_time: u64) -> MaResult<()> {
+    fn set_time(&self, local_time: u64) -> MaResult<()> {
         node_ffi::ma_node_set_time(self, local_time)
     }
 }
@@ -762,9 +752,9 @@ pub(super) mod node_ffi {
 
     #[inline]
     pub(crate) fn ma_node_attach_output_bus<P: AsNodePtr + ?Sized, Q: AsNodePtr + ?Sized>(
-        node: &mut P,
+        node: &P,
         output_bus_index: u32,
-        other_node: &mut Q,
+        other_node: &Q,
         other_node_input_bus_index: u32,
     ) -> MaResult<()> {
         unsafe {
@@ -780,7 +770,7 @@ pub(super) mod node_ffi {
 
     #[inline]
     pub(crate) fn ma_node_detach_output_bus<P: AsNodePtr + ?Sized>(
-        node: &mut P,
+        node: &P,
         output_bus_index: u32,
     ) -> MaResult<()> {
         let res = unsafe {
@@ -790,16 +780,14 @@ pub(super) mod node_ffi {
     }
 
     #[inline]
-    pub(crate) fn ma_node_detach_all_output_buses<P: AsNodePtr + ?Sized>(
-        node: &mut P,
-    ) -> MaResult<()> {
+    pub(crate) fn ma_node_detach_all_output_buses<P: AsNodePtr + ?Sized>(node: &P) -> MaResult<()> {
         let res = unsafe { sys::ma_node_detach_all_output_buses(private_node::node_ptr(node)) };
         MaudioError::check(res)
     }
 
     #[inline]
     pub(crate) fn ma_node_set_output_bus_volume<P: AsNodePtr + ?Sized>(
-        node: &mut P,
+        node: &P,
         output_bus_index: sys::ma_uint32,
         volume: f32,
     ) -> MaResult<()> {
@@ -815,7 +803,7 @@ pub(super) mod node_ffi {
 
     #[inline]
     pub(crate) fn ma_node_get_output_bus_volume<P: AsNodePtr + ?Sized>(
-        node: &mut P,
+        node: &P,
         output_bus_index: sys::ma_uint32,
     ) -> f32 {
         unsafe {
@@ -825,7 +813,7 @@ pub(super) mod node_ffi {
 
     #[inline]
     pub(crate) fn ma_node_set_state<P: AsNodePtr + ?Sized>(
-        node: &mut P,
+        node: &P,
         state: NodeState,
     ) -> MaResult<()> {
         let res = unsafe { sys::ma_node_set_state(private_node::node_ptr(node), state.into()) };
@@ -840,7 +828,7 @@ pub(super) mod node_ffi {
 
     #[inline]
     pub(crate) fn ma_node_set_state_time<P: AsNodePtr + ?Sized>(
-        node: &mut P,
+        node: &P,
         state: NodeState,
         global_time: u64,
     ) -> MaResult<()> {
@@ -891,7 +879,7 @@ pub(super) mod node_ffi {
 
     #[inline]
     pub(crate) fn ma_node_set_time<P: AsNodePtr + ?Sized>(
-        node: &mut P,
+        node: &P,
         local_time: u64,
     ) -> MaResult<()> {
         let res = unsafe { sys::ma_node_set_time(private_node::node_ptr(node), local_time) };